fix(structure): Move sw6-connector to packages/shopware-6-client/
