+++ packages/shopware-6-client/src/interfaces/models/framework/data-abstraction-layer/Translation.ts
@@ -1,6 +1,6 @@
 import { LanguageEntity } from "./context/LanguageEntity";
 
 export interface Translation {
-    languageId: string;
-    language: LanguageEntity;
+  languageId: string;
+  language: LanguageEntity;
 }
