+++ packages/shopware-6-client/src/interfaces/models/framework/language/LanguageCollection.ts
@@ -1,4 +1,4 @@
 import { Language } from "./Language";
 export interface LanguageCollection {
-    [index: number]: Language;
+  [index: number]: Language;
 }
