+++ packages/shopware-6-client/src/interfaces/models/framework/plugin/PluginTranslationCollection.ts
@@ -1,4 +1,4 @@
 import { PluginTranslation } from "./PluginTranslation";
 export interface PluginTranslationCollection {
-    [index: number]: PluginTranslation;
+  [index: number]: PluginTranslation;
 }
