+++ packages/shopware-6-client/src/interfaces/models/framework/pricing/ListingPrice.ts
@@ -1,0 +1,7 @@
+import { Price } from "./Price";
+export interface ListingPrice {
+  currencyId: number;
+  ruleId: string;
+  from: Price;
+  to: Price;
+}
