+++ packages/shopware-6-client/src/interfaces/models/framework/pricing/ListingPriceCollection.ts
@@ -1,5 +1,4 @@
 import { ListingPrice } from "./ListingPrice";
 export interface ListingPriceCollection {
-    [index: number]: ListingPrice;
+  [index: number]: ListingPrice;
 }
-
