+++ packages/shopware-6-client/src/interfaces/models/framework/pricing/Price.ts
@@ -1,0 +1,6 @@
+export interface Price {
+  currencyId: number;
+  net: number;
+  gross: number;
+  linked: boolean;
+}
