+++ packages/shopware-6-client/src/interfaces/models/framework/pricing/PriceCollection.ts
@@ -1,4 +1,4 @@
 import { Price } from "./Price";
 export interface PriceCollection {
-    [index: number]: Price;
+  [index: number]: Price;
 }
