+++ packages/shopware-6-client/src/interfaces/models/framework/rule/RuleCollection.ts
@@ -1,5 +1,5 @@
 import { Rule } from "./Rule";
 
 export interface RuleCollection {
-    [index: number]: Rule;
+  [index: number]: Rule;
 }
