+++ packages/shopware-6-client/src/interfaces/models/framework/snippet/SnippetCollection.ts
@@ -1,4 +1,4 @@
 import { SnippetEntity } from "./SnippetEntity";
 export interface SnippetCollection {
-    [index: number]: SnippetEntity;
+  [index: number]: SnippetEntity;
 }
