+++ packages/shopware-6-client/src/interfaces/models/framework/snippet/SnippetEntity.ts
@@ -1,0 +1,9 @@
+import { SnippetSetEntity } from "./SnippetSetEntity";
+export interface SnippetEntity {
+  setId: string;
+  translationKey: string;
+  value: string;
+  author: string;
+  set: SnippetSetEntity | null;
+  customFields: [] | null;
+}
