+++ packages/shopware-6-client/src/interfaces/models/framework/snippet/SnippetSetCollection.ts
@@ -1,4 +1,4 @@
 import { SnippetSetEntity } from "./SnippetSetEntity";
 interface SnippetSetCollection {
-    [index: number]: SnippetSetEntity;
+  [index: number]: SnippetSetEntity;
 }
