+++ packages/shopware-6-client/src/interfaces/models/framework/snippet/SnippetSetEntity.ts
@@ -1,10 +1,10 @@
 import { SnippetCollection } from "./SnippetCollection";
 import { SalesChannelDomainCollection } from "../state-machine/SalesChannelDomainCollection";
 export interface SnippetSetEntity {
-    name: string;
-    baseFile: string;
-    iso: string;
-    snippets: SnippetCollection | null;
-    salesChannelDomains: SalesChannelDomainCollection | null;
-    customFields: [] | null;
+  name: string;
+  baseFile: string;
+  iso: string;
+  snippets: SnippetCollection | null;
+  salesChannelDomains: SalesChannelDomainCollection | null;
+  customFields: [] | null;
 }
