+++ packages/shopware-6-client/src/interfaces/models/framework/struct/Collection.ts
@@ -1,3 +1,3 @@
 export interface Collection {
-    elements: [];
+  elements: [];
 }
