+++ packages/shopware-6-client/src/interfaces/models/framework/struct/Entity.ts
@@ -1,0 +1,7 @@
+export interface Entity {
+  _uniqueIdentifier: string;
+  versionId: string;
+  translated: [];
+  createdAt: Date | null;
+  updatedAt: Date | null;
+}
