+++ packages/shopware-6-client/src/interfaces/models/framework/struct/EntityCollection.ts
@@ -1,4 +1,4 @@
 import { Entity } from "./Entity";
 export interface EntityCollection {
-    [index: number]: Entity;
+  [index: number]: Entity;
 }
