--- sw6-connector/interfaces/models/framework/plugin/PluginTranslation.ts
@@ -1,11 +1,0 @@
-import { Plugin } from "./Plugin";
-export interface PluginTranslation {
-    pluginId: string;
-    label: string | null;
-    description: string | null;
-    manufacturerLink: string | null;
-    supportLink: string | null;
-    changelog: [] | null;
-    plugin: Plugin | null;
-    customFields: [] | null;
-}
