--- sw6-connector/interfaces/models/framework/pricing/ListingPrice.ts
@@ -1,7 +1,0 @@
-import { Price } from "./Price";
-export interface ListingPrice {
-    currencyId: number;
-    ruleId: string;
-    from: Price;
-    to: Price;
-}
