--- sw6-connector/interfaces/models/framework/pricing/Price.ts
@@ -1,6 +1,0 @@
-export interface Price {
-    currencyId: number;
-    net: number;
-    gross: number;
-    linked: boolean;
-}
