--- sw6-connector/interfaces/models/framework/snippet/SnippetEntity.ts
@@ -1,9 +1,0 @@
-import { SnippetSetEntity } from "./SnippetSetEntity";
-export interface SnippetEntity {
-    setId: string;
-    translationKey: string;
-    value: string;
-    author: string;
-    set: SnippetSetEntity | null;
-    customFields: [] | null;
-}
