--- sw6-connector/interfaces/models/framework/struct/Entity.ts
@@ -1,8 +1,0 @@
-export interface Entity {
-    _uniqueIdentifier: string;
-    versionId: string;
-    translated: [];
-    createdAt: Date | null;
-    updatedAt: Date | null;
-  }
-
